Refactoring XmlConverter.
