+++ src/test/scala/advxml/syntax/XmlConverterSyntaxTest.scala
@@ -1,6 +1,7 @@
 package advxml.syntax
 
-import advxml.core.convert.{ModelToXml, ValidatedConverter, XmlToModel}
+import advxml.core.convert.ValidatedConverter
+import advxml.core.convert.xml.{ModelToXml, XmlToModel}
 import advxml.core.validate.ValidatedEx
 import cats.data.Validated.Valid
 import org.scalatest.FunSuite
